Cleanup and converge V1 and V2 SchemaConformingTransformer
